+++ src/common.rs
@@ -63,8 +63,11 @@ pub trait Component {
     /// returns the (id, Ports) of the component
     fn get_id_ports(&self) -> (Id, Ports);
 
-    /// evaluation function
-    fn evaluate(&self, _simulator: &mut Simulator) {}
+    /// evaluate component based on current internal state
+    fn clock(&self, _simulator: &mut Simulator) {}
+
+    /// update component internal state
+    fn un_clock(&self) {}
 }
 
 // Specific functionality for Vizia frontend