+++ src/gui_vizia/gui.rs
@@ -167,6 +167,7 @@ pub fn gui(cs: &ComponentStore, path: &PathBuf) {
                                 let oc = wrapper_oc.get(cx);
                                 for (i, c) in oc.iter().enumerate() {
                                     VStack::new(cx, |cx| {
+                                        // left pane bar
                                         HStack::new(cx, move |cx| {
                                             Button::new(
                                                 cx,
@@ -244,37 +245,44 @@ pub fn gui(cs: &ComponentStore, path: &PathBuf) {
                     );
                 });
 
-                // Grid area
-                Grid::new(cx, |cx| {
-                    // (re-)bind all components when simulator changed
-                    Binding::new(
-                        cx,
-                        GuiData::simulator.then(Simulator::ordered_components),
-                        |cx, wrapper_oc| {
-                            VStack::new(cx, |cx| {
-                                let oc = wrapper_oc.get(cx);
-                                for (i, c) in oc.iter().enumerate() {
-                                    error!("comp id {}", i);
-                                    VStack::new(cx, |cx| {
-                                        c.view(cx);
-                                    })
-                                    .position_type(PositionType::SelfDirected)
-                                    .size(Auto)
-                                    .on_mouse_down(
-                                        move |ex, button| {
-                                            if button == MouseButton::Right {
-                                                error!("on_mouse_down {:?}", i);
-                                                ex.emit(GuiEvent::ShowLeftPanel(i))
-                                            }
-                                        },
-                                    );
-                                }
-                            })
-                            .border_color(Color::black())
-                            .border_width(Pixels(1.0));
-                        },
-                    )
-                });
+                ScrollView::new(cx, 0.0, 0.0, true, true, |cx| {
+                    // Grid area
+                    Grid::new(cx, |cx| {
+                        // (re-)bind all components when simulator changed
+                        Binding::new(
+                            cx,
+                            GuiData::simulator.then(Simulator::ordered_components),
+                            |cx, wrapper_oc| {
+                                VStack::new(cx, |cx| {
+                                    let oc = wrapper_oc.get(cx);
+                                    for (i, c) in oc.iter().enumerate() {
+                                        error!("comp id {}", i);
+                                        VStack::new(cx, |cx| {
+                                            c.view(cx);
+                                        })
+                                        .position_type(PositionType::SelfDirected)
+                                        .size(Auto)
+                                        .on_mouse_down(
+                                            move |ex, button| {
+                                                if button == MouseButton::Right {
+                                                    error!("on_mouse_down {:?}", i);
+                                                    ex.emit(GuiEvent::ShowLeftPanel(i))
+                                                }
+                                            },
+                                        );
+                                    }
+                                })
+                                .border_color(Color::black())
+                                .border_width(Pixels(1.0))
+                                .overflow(Overflow::Hidden);
+                            },
+                        )
+                    })
+                    .height(Pixels(1080.0))
+                    .width(Pixels(1920.0));
+                })
+                // .size(Units::Pixels(300.0))
+                .class("bg-default");
 
                 // Right pane
                 Label::new(cx, "Right").top(Pixels(0.0));