+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Tracking changes per date:
 
+## 230719
+
+- `ProbeEdit`, a component for interactive debugging (and maybe some end usage as well). It allows you to enter a value (dec/hex) for a signal. It acts as a register so its content will be used in the next clock cycle (one could also think of changing this to act directly by triggering some re-evaluation, not sure). `ProbeEdit` also implements a proper history buffer so you can reverse the simulation.
+
+- `un_clock` method in the `Component` trait. It does not alter the simulation state (for that we already have a history buffer), components with internal state (e.g., the `ProbeEdit` component) implements `un_clock` and keep a local history. Maybe we will have some helper functions to that end.
+
 ## 230718
 
 - Panes, right click component to display component interior on left panel. Close left panel view by the X (Close) button.